+++ packages/isaacscript-common/src/classes/features/other/extraConsoleCommands/commands.ts
@@ -111,11 +111,7 @@ import {
   asCollectibleType,
   asTrinketType,
 } from "../../../../functions/types";
-import {
-  iRange,
-  printConsole,
-  printEnabled,
-} from "../../../../functions/utils";
+import { iRange, printEnabled } from "../../../../functions/utils";
 import { CARD_NAME_TO_TYPE_MAP } from "../../../../maps/cardNameToTypeMap";
 import { CHARACTER_NAME_TO_TYPE_MAP } from "../../../../maps/characterNameToTypeMap";
 import { COLLECTIBLE_NAME_TO_TYPE_MAP } from "../../../../maps/collectibleNameToTypeMap";
@@ -141,7 +137,7 @@ import { v } from "./v";
  */
 export function addCharges(params: string): void {
   if (params === "") {
-    printConsole(
+    print(
       "You must specify a slot number. (Use 0 for the primary slot, 1 for the Schoolbag slot, 2 for the pocket item slot, and 3 for the Dice Bag slot.)",
     );
     return;
@@ -150,7 +146,7 @@ export function addCharges(params: string): void {
   const args = params.split(" ");
 
   if (args.length !== 1 && args.length !== 2) {
-    printConsole("That is an invalid amount of arguments.");
+    print("That is an invalid amount of arguments.");
     return;
   }
 
@@ -158,7 +154,7 @@ export function addCharges(params: string): void {
 
   const activeSlot = tonumber(activeSlotString) as ActiveSlot | undefined;
   if (activeSlot === undefined) {
-    printConsole(`The provided slot number is invalid: ${activeSlotString}`);
+    print(`The provided slot number is invalid: ${activeSlotString}`);
     return;
   }
 
@@ -166,7 +162,7 @@ export function addCharges(params: string): void {
     activeSlot < ActiveSlot.PRIMARY ||
     activeSlot > ActiveSlot.POCKET_SINGLE_USE
   ) {
-    printConsole(`The provided slot number is invalid: ${activeSlot}`);
+    print(`The provided slot number is invalid: ${activeSlot}`);
     return;
   }
 
@@ -174,7 +170,7 @@ export function addCharges(params: string): void {
   if (numChargeString !== undefined) {
     const numChargesAttempt = tonumber(numChargeString);
     if (numChargesAttempt === undefined) {
-      printConsole(`The provided charge amount is invalid: ${numChargeString}`);
+      print(`The provided charge amount is invalid: ${numChargeString}`);
       return;
     }
     numCharges = numChargesAttempt;
@@ -197,7 +193,7 @@ export function ascent(): void {
   game.SetStateFlag(GameStateFlag.BACKWARDS_PATH_INIT, true);
   game.SetStateFlag(GameStateFlag.BACKWARDS_PATH, true);
 
-  printConsole("Set Ascent flags.");
+  print("Set Ascent flags.");
 }
 
 /** Warps to the first Clean Bedroom or Dirty Bedroom on the floor. */
@@ -218,7 +214,7 @@ export function bedroom(): void {
     return;
   }
 
-  printConsole("There are no Clean Bedrooms or Dirty Bedrooms on this floor.");
+  print("There are no Clean Bedrooms or Dirty Bedrooms on this floor.");
 }
 
 /**
@@ -249,7 +245,7 @@ export function bloodCharges(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of charges to add.");
+      print("That is an invalid amount of charges to add.");
       return;
     }
 
@@ -274,7 +270,7 @@ export function bomb(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of bombs to add.");
+      print("That is an invalid amount of bombs to add.");
       return;
     }
 
@@ -294,7 +290,7 @@ export function bombs(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of bombs to add.");
+      print("That is an invalid amount of bombs to add.");
       return;
     }
 
@@ -350,7 +346,7 @@ export function brokenHearts(params: string): void {
  */
 export function card(params: string): void {
   if (params === "") {
-    printConsole("You must specify a card name or number.");
+    print("You must specify a card name or number.");
     return;
   }
 
@@ -359,14 +355,14 @@ export function card(params: string): void {
   if (num === undefined) {
     const match = getMapPartialMatch(params, CARD_NAME_TO_TYPE_MAP);
     if (match === undefined) {
-      printConsole(`Unknown card: ${params}`);
+      print(`Unknown card: ${params}`);
       return;
     }
 
     cardType = match[1];
   } else {
     if (num < FIRST_CARD_TYPE || num > LAST_VANILLA_CARD_TYPE) {
-      printConsole(`Invalid card sub-type: ${num}`);
+      print(`Invalid card sub-type: ${num}`);
       return;
     }
 
@@ -375,7 +371,7 @@ export function card(params: string): void {
 
   const cardName = getCardName(cardType);
   Isaac.ExecuteCommand(`g k${cardType}`);
-  printConsole(`Gave card: ${cardName} (${cardType})`);
+  print(`Gave card: ${cardName} (${cardType})`);
 }
 
 /** Spawns every card on the ground, starting at the top-left-most tile. */
@@ -418,7 +414,7 @@ export function chaosCardTears(): void {
  */
 export function character(params: string): void {
   if (params === "") {
-    printConsole("You must specify a character name or number.");
+    print("You must specify a character name or number.");
     return;
   }
 
@@ -427,14 +423,14 @@ export function character(params: string): void {
   if (num === undefined) {
     const match = getMapPartialMatch(params, CHARACTER_NAME_TO_TYPE_MAP);
     if (match === undefined) {
-      printConsole(`Unknown character: ${params}`);
+      print(`Unknown character: ${params}`);
       return;
     }
 
     playerType = match[1];
   } else {
     if (num < FIRST_CHARACTER || num > LAST_VANILLA_CHARACTER) {
-      printConsole(`Invalid player sub-type: ${num}`);
+      print(`Invalid player sub-type: ${num}`);
       return;
     }
 
@@ -443,7 +439,7 @@ export function character(params: string): void {
 
   const characterName = getCharacterName(playerType);
   restart(playerType);
-  printConsole(`Restarting as character: ${characterName} (${playerType})`);
+  print(`Restarting as character: ${characterName} (${playerType})`);
 }
 
 /** Alias for the "addCharges" command. */
@@ -465,7 +461,7 @@ export function coin(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of coins to add.");
+      print("That is an invalid amount of coins to add.");
       return;
     }
 
@@ -485,7 +481,7 @@ export function coins(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of coins to add.");
+      print("That is an invalid amount of coins to add.");
       return;
     }
 
@@ -534,7 +530,7 @@ export function damage(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole(`The provided damage amount is invalid: ${params}`);
+      print(`The provided damage amount is invalid: ${params}`);
       return;
     }
 
@@ -600,7 +596,7 @@ export function dungeon(): void {
 export function effects(): void {
   const player = Isaac.GetPlayer();
   logPlayerEffects(player);
-  printConsole('Logged the player\'s effects to the "log.txt" file.');
+  print('Logged the player\'s effects to the "log.txt" file.');
 }
 
 /** Alias for the "iAmError" command. */
@@ -657,7 +653,7 @@ export function getChallenge(): void {
     challengeName === undefined
       ? `${challenge} (custom)`
       : `Challenge.${challengeName} (${challenge})`;
-  printConsole(`The current challenge is: ${challengeDescription}`);
+  print(`The current challenge is: ${challengeDescription}`);
 }
 
 /** Prints the charge for the specified slot. By default, will use `ActiveSlot.PRIMARY`. */
@@ -666,7 +662,7 @@ export function getCharge(params: string): void {
   if (params !== "") {
     const num = tonumber(params) as ActiveSlot | undefined;
     if (num === undefined) {
-      printConsole(`The provided slot number is invalid: ${params}`);
+      print(`The provided slot number is invalid: ${params}`);
       return;
     }
 
@@ -675,7 +671,7 @@ export function getCharge(params: string): void {
 
   const player = Isaac.GetPlayer();
   const totalCharge = getTotalCharge(player, activeSlot);
-  printConsole(
+  print(
     `Total charge for ActiveSlot.${ActiveSlot[activeSlot]} (${activeSlot}) is: ${totalCharge}`,
   );
 }
@@ -684,7 +680,7 @@ export function getCharge(params: string): void {
 export function getPosition(): void {
   for (const player of getPlayers()) {
     const playerName = getPlayerName(player);
-    printConsole(
+    print(
       `Player position for ${playerName}: (${player.Position.X}, ${player.Position.Y})`,
     );
   }
@@ -705,7 +701,7 @@ export function gigaBomb(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of Giga Bombs to add.");
+      print("That is an invalid amount of Giga Bombs to add.");
       return;
     }
 
@@ -821,7 +817,7 @@ export function key(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of keys to add.");
+      print("That is an invalid amount of keys to add.");
       return;
     }
 
@@ -841,7 +837,7 @@ export function keys(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of keys to add.");
+      print("That is an invalid amount of keys to add.");
       return;
     }
 
@@ -976,7 +972,7 @@ export function oneHP(): void {
     npc.HitPoints = 1;
   }
 
-  printConsole("Set every NPC to 1 HP.");
+  print("Set every NPC to 1 HP.");
 }
 
 /**
@@ -990,7 +986,7 @@ export function oneHP(): void {
  */
 export function pill(params: string): void {
   if (params === "") {
-    printConsole("You must specify a pill name or number.");
+    print("You must specify a pill name or number.");
     return;
   }
 
@@ -999,14 +995,14 @@ export function pill(params: string): void {
   if (num === undefined) {
     const match = getMapPartialMatch(params, PILL_NAME_TO_EFFECT_MAP);
     if (match === undefined) {
-      printConsole(`Unknown pill effect: ${params}`);
+      print(`Unknown pill effect: ${params}`);
       return;
     }
 
     pillEffect = match[1];
   } else {
     if (num < FIRST_PILL_EFFECT || num > LAST_VANILLA_PILL_EFFECT) {
-      printConsole(`Invalid pill effect ID: ${num}`);
+      print(`Invalid pill effect ID: ${num}`);
       return;
     }
 
@@ -1015,7 +1011,7 @@ export function pill(params: string): void {
 
   const pillEffectName = getPillEffectName(pillEffect);
   Isaac.ExecuteCommand(`g p${pillEffect}`);
-  printConsole(`Gave pill: ${pillEffectName} (${pillEffect})`);
+  print(`Gave pill: ${pillEffectName} (${pillEffect})`);
 }
 
 /** Spawns every pill on the ground, starting at the top-left-most tile. */
@@ -1047,20 +1043,18 @@ export function playSound(params: string): void {
 /** Sets the player's pocket item to the specified collectible type. */
 export function pocket(params: string): void {
   if (params === "") {
-    printConsole(
-      "You must supply a collectible type to put as the pocket item.",
-    );
+    print("You must supply a collectible type to put as the pocket item.");
     return;
   }
 
   const collectibleType = tonumber(params) as CollectibleType | undefined;
   if (collectibleType === undefined) {
-    printConsole("That is an invalid collectible type.");
+    print("That is an invalid collectible type.");
     return;
   }
 
   if (!isValidCollectibleType(collectibleType)) {
-    printConsole("That is an invalid collectible type.");
+    print("That is an invalid collectible type.");
     return;
   }
 
@@ -1077,7 +1071,7 @@ export function poopMana(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of mana to add.");
+      print("That is an invalid amount of mana to add.");
       return;
     }
 
@@ -1111,7 +1105,7 @@ export function right(params: string): void {
 /** Logs information about the room to the "log.txt" file. */
 export function room(): void {
   logRoom();
-  printConsole('Logged room information to the "log.txt" file.');
+  print('Logged room information to the "log.txt" file.');
 }
 
 /**
@@ -1143,7 +1137,7 @@ export function runTests(): void {
  */
 export function s(params: string): void {
   if (params === "") {
-    printConsole("You must specify a stage number.");
+    print("You must specify a stage number.");
     return;
   }
 
@@ -1167,12 +1161,12 @@ export function s(params: string): void {
 
   const stage = tonumber(stageString) as LevelStage | undefined;
   if (stage === undefined) {
-    printConsole(`That is an invalid stage number: ${stage}`);
+    print(`That is an invalid stage number: ${stage}`);
     return;
   }
 
   if (stage < FIRST_STAGE || stage > LAST_STAGE) {
-    printConsole(
+    print(
       `Invalid stage number; must be between ${FIRST_STAGE} and ${LAST_STAGE}.`,
     );
     return;
@@ -1206,7 +1200,7 @@ export function seedStick(): void {
 /** Logs all of the current run's seed effects to the "log.txt" file. */
 export function seeds(): void {
   logSeedEffects();
-  printConsole('Logged the seed effects to the "log.txt" file.');
+  print('Logged the seed effects to the "log.txt" file.');
 }
 
 /**
@@ -1215,7 +1209,7 @@ export function seeds(): void {
  */
 export function setCharges(params: string): void {
   if (params === "") {
-    printConsole(
+    print(
       "You must specify a slot number and a charge amount. (Use 0 for the primary slot, 1 for the Schoolbag slot, 2 for the pocket item slot, and 3 for the Dice Bag slot.)",
     );
     return;
@@ -1224,12 +1218,12 @@ export function setCharges(params: string): void {
   const args = params.split(" ");
 
   if (args.length === 1) {
-    printConsole("You must specify the amount of charge to set.");
+    print("You must specify the amount of charge to set.");
     return;
   }
 
   if (args.length !== 2) {
-    printConsole("That is an invalid amount of arguments.");
+    print("That is an invalid amount of arguments.");
     return;
   }
 
@@ -1237,24 +1231,24 @@ export function setCharges(params: string): void {
 
   const activeSlot = tonumber(activeSlotString) as ActiveSlot | undefined;
   if (activeSlot === undefined) {
-    printConsole(`The provided slot number is invalid: ${activeSlotString}`);
+    print(`The provided slot number is invalid: ${activeSlotString}`);
     return;
   }
 
   const activeSlots = getEnumValues(ActiveSlot);
   if (!activeSlots.includes(activeSlot)) {
-    printConsole(`The provided slot number is invalid: ${activeSlot}`);
+    print(`The provided slot number is invalid: ${activeSlot}`);
     return;
   }
 
   const chargeNum = tonumber(chargeString);
   if (chargeNum === undefined) {
-    printConsole(`The provided charge amount is invalid: ${chargeString}`);
+    print(`The provided charge amount is invalid: ${chargeString}`);
     return;
   }
 
   if (chargeNum < 0) {
-    printConsole(`The provided charge amount is invalid: ${chargeNum}`);
+    print(`The provided charge amount is invalid: ${chargeNum}`);
     return;
   }
 
@@ -1270,13 +1264,13 @@ export function setCharges(params: string): void {
  */
 export function setPosition(params: string): void {
   if (params === "") {
-    printConsole('You must specify a position. (e.g. "setPosition 100 50")');
+    print('You must specify a position. (e.g. "setPosition 100 50")');
     return;
   }
 
   const args = params.split(" ");
   if (args.length !== 2) {
-    printConsole('You must specify a position. (e.g. "setPosition 100 50")');
+    print('You must specify a position. (e.g. "setPosition 100 50")');
     return;
   }
 
@@ -1284,13 +1278,13 @@ export function setPosition(params: string): void {
 
   const x = tonumber(xString);
   if (x === undefined) {
-    printConsole(`That is an invalid x value: ${xString}`);
+    print(`That is an invalid x value: ${xString}`);
     return;
   }
 
   const y = tonumber(yString);
   if (y === undefined) {
-    printConsole(`That is an invalid y value: ${yString}`);
+    print(`That is an invalid y value: ${yString}`);
     return;
   }
 
@@ -1319,7 +1313,7 @@ export function soulCharges(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole("That is an invalid amount of charges to add.");
+      print("That is an invalid amount of charges to add.");
       return;
     }
 
@@ -1347,7 +1341,7 @@ export function soulHearts(params: string): void {
 export function sound(params: string): void {
   const soundEffect = tonumber(params) as SoundEffect | undefined;
   if (soundEffect === undefined) {
-    printConsole("That is an invalid sound effect ID.");
+    print("That is an invalid sound effect ID.");
     return;
   }
 
@@ -1357,9 +1351,7 @@ export function sound(params: string): void {
 /** Logs all of the currently playing sound effects to the "log.txt" file. */
 export function sounds(): void {
   logSounds();
-  printConsole(
-    'Logged the currently playing sound effects to the "log.txt" file.',
-  );
+  print('Logged the currently playing sound effects to the "log.txt" file.');
 }
 
 /**
@@ -1373,7 +1365,7 @@ export function spam(): void {
 
 export function spawnCollectible(params: string): void {
   if (params === "") {
-    printConsole(
+    print(
       "You must specify the name or number corresponding to the collectible type.",
     );
     return;
@@ -1384,7 +1376,7 @@ export function spawnCollectible(params: string): void {
   if (collectibleTypeNumber === undefined) {
     const match = getMapPartialMatch(params, COLLECTIBLE_NAME_TO_TYPE_MAP);
     if (match === undefined) {
-      printConsole(`Unknown collectible: ${params}`);
+      print(`Unknown collectible: ${params}`);
       return;
     }
 
@@ -1401,7 +1393,7 @@ export function spawnCollectible(params: string): void {
 /** Spawns a golden version of the specified trinket type. */
 export function spawnGoldenTrinket(params: string): void {
   if (params === "") {
-    printConsole(
+    print(
       "You must specify the name or number corresponding to the trinket type.",
     );
     return;
@@ -1412,7 +1404,7 @@ export function spawnGoldenTrinket(params: string): void {
   if (trinketTypeNumber === undefined) {
     const match = getMapPartialMatch(params, TRINKET_NAME_TO_TYPE_MAP);
     if (match === undefined) {
-      printConsole(`Unknown trinket: ${params}`);
+      print(`Unknown trinket: ${params}`);
       return;
     }
 
@@ -1429,7 +1421,7 @@ export function spawnGoldenTrinket(params: string): void {
 
 export function spawnTrinket(params: string): void {
   if (params === "") {
-    printConsole(
+    print(
       "You must specify the name or number corresponding to the trinket type.",
     );
     return;
@@ -1440,7 +1432,7 @@ export function spawnTrinket(params: string): void {
   if (trinketTypeNumber === undefined) {
     const match = getMapPartialMatch(params, TRINKET_NAME_TO_TYPE_MAP);
     if (match === undefined) {
-      printConsole(`Unknown trinket: ${params}`);
+      print(`Unknown trinket: ${params}`);
       return;
     }
 
@@ -1465,7 +1457,7 @@ export function speed(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole(`The provided speed amount is invalid: ${params}`);
+      print(`The provided speed amount is invalid: ${params}`);
       return;
     }
 
@@ -1509,7 +1501,7 @@ export function tears(params: string): void {
   if (params !== "") {
     const num = tonumber(params);
     if (num === undefined) {
-      printConsole(`The provided tear delay amount is invalid: ${params}`);
+      print(`The provided tear delay amount is invalid: ${params}`);
       return;
     }
 
@@ -1554,7 +1546,7 @@ export function unknown(): void {
 /** If currently on a set seed, changes to an unseeded state and restarts the game. */
 export function unseed(): void {
   if (!onSetSeed()) {
-    printConsole("You are not on a set seed, so you cannot unseed the run.");
+    print("You are not on a set seed, so you cannot unseed the run.");
     return;
   }
 
@@ -1577,7 +1569,7 @@ export function up(params: string): void {
  */
 export function warp(params: string): void {
   if (params === "") {
-    printConsole("You must specify a room type name or number.");
+    print("You must specify a room type name or number.");
     return;
   }
 
@@ -1586,14 +1578,14 @@ export function warp(params: string): void {
   if (num === undefined) {
     const match = getMapPartialMatch(params, ROOM_NAME_TO_TYPE_MAP);
     if (match === undefined) {
-      printConsole(`Unknown room type: ${params}`);
+      print(`Unknown room type: ${params}`);
       return;
     }
 
     roomType = match[1];
   } else {
     if (num < FIRST_ROOM_TYPE || num > LAST_ROOM_TYPE) {
-      printConsole(`Invalid room type: ${num}`);
+      print(`Invalid room type: ${num}`);
       return;
     }