+++ packages/isaacscript-common/src/functions/deepCopyTests.ts
@@ -6,7 +6,6 @@ import { deepCopy } from "./deepCopy";
 import { log } from "./log";
 import { isDefaultMap, isTSTLMap, isTSTLSet } from "./tstlClass";
 import { isNumber, isString, isTable } from "./types";
-import { printConsole } from "./utils";
 
 /**
  * Run the suite of tests that prove that the "deepCopy" helper function works properly.
@@ -41,7 +40,7 @@ export function runDeepCopyTests(): void {
 
   const successText = "All deep copy tests passed!";
   log(successText);
-  printConsole(successText);
+  print(successText);
 }
 
 function copiedObjectIsTable() {